+++ packages/cheqd-blockchain-modules/tests/accumulator-module.test.js
@@ -1,4 +1,0 @@
-import generateAccumulatorModuleTests from "@docknetwork/credential-sdk/modules/tests/accumulator-module";
-import { tests } from "./common";
-
-tests("AccumulatorModule", generateAccumulatorModuleTests);