+++ packages/cheqd-blockchain-modules/tests/attest-module.test.js
@@ -1,4 +1,0 @@
-import generateAttestModuleTests from "@docknetwork/credential-sdk/modules/tests/attest-module";
-import { tests } from "./common";
-
-tests("AttestModule", generateAttestModuleTests);