+++ packages/cheqd-blockchain-modules/tests/blob-module.test.js
@@ -1,4 +1,0 @@
-import generateBlobModuleTests from "@docknetwork/credential-sdk/modules/tests/blob-module";
-import { tests } from "./common";
-
-tests("BlobModule", generateBlobModuleTests);