+++ packages/cheqd-blockchain-modules/tests/did-module.test.js
@@ -1,5 +1,0 @@
-import generateDIDModuleTests from "@docknetwork/credential-sdk/modules/tests/did-module";
-import { CheqdParamsId } from "@docknetwork/credential-sdk/types";
-import { tests } from "./common";
-
-tests("DIDModule", generateDIDModuleTests, { ParamsId: CheqdParamsId });