+++ packages/cheqd-blockchain-modules/tests/offchain-signatures-module.test.js
@@ -1,4 +1,0 @@
-import generateOffchainSignatureModuleTests from "@docknetwork/credential-sdk/modules/tests/offchain-signatures-module";
-import { tests } from "./common";
-
-tests("OffchainSignaturesModule", generateOffchainSignatureModuleTests);