+++ packages/cheqd-blockchain-modules/tests/status-list-credential-module.test.js
@@ -1,4 +1,0 @@
-import generateStatusListCredentialModuleTests from "@docknetwork/credential-sdk/modules/tests/status-list-credential-module";
-import { tests } from "./common";
-
-tests("StatusListCredentialModule", generateStatusListCredentialModuleTests);