+++ packages/credential-sdk/src/modules/tests/did-module.js
@@ -1,5 +1,5 @@
-import { DidKeypair, Ed25519Keypair } from "../../keypairs";
-import { DIDResolver } from "../../resolver";
+import { DidKeypair, Ed25519Keypair } from '../../keypairs';
+import { DIDResolver } from '../../resolver';
 import {
   DIDDocument,
   BBSPublicKeyValue,
@@ -8,21 +8,21 @@ import {
   PSPublicKeyValue,
   VerificationMethodRef,
   ServiceEndpoint,
-} from "../../types";
-import { TypedBytes } from "../../types/generic";
-import { NoDIDError } from "../abstract/did/errors";
-import { testIf } from "./common";
+} from '../../types';
+import { TypedBytes } from '../../types/generic';
+import { NoDIDError } from '../abstract/did/errors';
+import { testIf } from './common';
 
 // eslint-disable-next-line jest/no-export
 export default function generateDIDModuleTests(
   { did: module },
   { Did, ParamsId },
-  filter = () => true
+  filter = () => true,
 ) {
   const test = testIf(filter);
 
   describe(`Using ${module.constructor.name} with ${Did.name}`, () => {
-    it.only("Creates basic `DIDDocument` with keys", async () => {
+    test('Creates basic `DIDDocument` with keys', async () => {
       const did = Did.random();
 
       const keyPair = Ed25519Keypair.random();
@@ -35,7 +35,7 @@ export default function generateDIDModuleTests(
       expect((await module.getDocument(did)).eq(document)).toBe(true);
     });
 
-    test("Creates `DIDDocument` containing BBS/BBSPlus/PS keys", async () => {
+    test('Creates `DIDDocument` containing BBS/BBSPlus/PS keys', async () => {
       const did = Did.random();
 
       const keyPair = Ed25519Keypair.random();
@@ -46,7 +46,7 @@ export default function generateDIDModuleTests(
       const bbsKey = new BBSPublicKeyValue(TypedBytes.random(100), paramsRef);
       const bbsPlusKey = new BBSPlusPublicKeyValue(
         TypedBytes.random(100),
-        paramsRef
+        paramsRef,
       );
       const psKey = new PSPublicKeyValue(TypedBytes.random(1000), paramsRef);
 
@@ -62,7 +62,7 @@ export default function generateDIDModuleTests(
       expect((await module.getDocument(did)).eq(document)).toBe(true);
     });
 
-    test("Updates `DIDDocument` containing BBS/BBSPlus/PS keys", async () => {
+    test('Updates `DIDDocument` containing BBS/BBSPlus/PS keys', async () => {
       const did = Did.random();
 
       const keyPair = Ed25519Keypair.random();
@@ -70,7 +70,7 @@ export default function generateDIDModuleTests(
 
       const bbsKey = new DidKey(new BBSPublicKeyValue(TypedBytes.random(100)));
       const bbsPlusKey = new DidKey(
-        new BBSPlusPublicKeyValue(TypedBytes.random(100))
+        new BBSPlusPublicKeyValue(TypedBytes.random(100)),
       );
       const psKey = new DidKey(new PSPublicKeyValue(TypedBytes.random(1000)));
 
@@ -99,17 +99,17 @@ export default function generateDIDModuleTests(
       ]);
     });
 
-    test("Creates `DIDDocument` containing services", async () => {
+    test('Creates `DIDDocument` containing services', async () => {
       const did = Did.random();
 
       const keyPair = Ed25519Keypair.random();
       const didKeypair = new DidKeypair([did, 1], keyPair);
 
-      const service1 = new ServiceEndpoint("LinkedDomains", [
-        "ServiceEndpoint#1",
+      const service1 = new ServiceEndpoint('LinkedDomains', [
+        'ServiceEndpoint#1',
       ]);
-      const service2 = new ServiceEndpoint("LinkedDomains", [
-        "ServiceEndpoint#2",
+      const service2 = new ServiceEndpoint('LinkedDomains', [
+        'ServiceEndpoint#2',
       ]);
 
       const document = DIDDocument.create(did, [didKeypair.didKey()], [], {
@@ -122,17 +122,17 @@ export default function generateDIDModuleTests(
       expect((await module.getDocument(did)).eq(document)).toBe(true);
     });
 
-    test("Updates `DIDDocument` containing services", async () => {
+    test('Updates `DIDDocument` containing services', async () => {
       const did = Did.random();
 
       const keyPair = Ed25519Keypair.random();
       const didKeypair = new DidKeypair([did, 1], keyPair);
 
-      const service1 = new ServiceEndpoint("LinkedDomains", [
-        "ServiceEndpoint#1",
+      const service1 = new ServiceEndpoint('LinkedDomains', [
+        'ServiceEndpoint#1',
       ]);
-      const service2 = new ServiceEndpoint("LinkedDomains", [
-        "ServiceEndpoint#2",
+      const service2 = new ServiceEndpoint('LinkedDomains', [
+        'ServiceEndpoint#2',
       ]);
 
       const document = DIDDocument.create(did, [didKeypair.didKey()]);
@@ -141,14 +141,14 @@ export default function generateDIDModuleTests(
 
       expect((await module.getDocument(did)).eq(document)).toBe(true);
 
-      document.addServiceEndpoint([did, "service1"], service1);
-      document.addServiceEndpoint([did, "service2"], service2);
+      document.addServiceEndpoint([did, 'service1'], service1);
+      document.addServiceEndpoint([did, 'service2'], service2);
 
       await module.updateDocument(document, didKeypair);
 
       expect((await module.getDocument(did)).eq(document)).toBe(true);
 
-      document.removeServiceEndpoint([did, "service2"]);
+      document.removeServiceEndpoint([did, 'service2']);
 
       await module.updateDocument(document, didKeypair);
 
@@ -181,15 +181,13 @@ export default function generateDIDModuleTests(
       expect((await module.getDocument(did)).eq(document)).toBe(true);
 
       document.addKey([did, 3], didKeypair3.didKey());
-      await expect(() =>
-        module.updateDocument(document, didKeypair2)
-      ).rejects.toThrow();
+      await expect(() => module.updateDocument(document, didKeypair2)).rejects.toThrow();
 
       await module.updateDocument(document, didKeypair1);
       expect((await module.getDocument(did)).eq(document)).toBe(true);
     });
 
-    test("Removes (deactivates) `DIDDocument`", async () => {
+    test('Removes (deactivates) `DIDDocument`', async () => {
       const did = Did.random();
 
       const keyPair = Ed25519Keypair.random();
@@ -202,7 +200,7 @@ export default function generateDIDModuleTests(
 
       await module.removeDocument(did, didKeypair);
       await expect(() => module.getDocument(did)).rejects.toThrow(
-        new NoDIDError(did)
+        new NoDIDError(did),
       );
     });
 
@@ -210,11 +208,11 @@ export default function generateDIDModuleTests(
       const did = Did.random();
 
       await expect(() => module.getDocument(did)).rejects.toThrow(
-        new NoDIDError(did)
+        new NoDIDError(did),
       );
     });
 
-    test("`DIDResolver`", async () => {
+    test('`DIDResolver`', async () => {
       const resolver = new DIDResolver(module);
       const did = Did.random();
 
@@ -226,7 +224,7 @@ export default function generateDIDModuleTests(
       await module.createDocument(document, didKeypair);
 
       expect(
-        DIDDocument.from(await resolver.resolve(String(did))).eq(document)
+        DIDDocument.from(await resolver.resolve(String(did))).eq(document),
       ).toBe(true);
     });
   });