+++ packages/credential-sdk/src/modules/tests/index.js
@@ -1,0 +1,6 @@
+export { default as generateAccumulatorModuleTests } from './accumulator-module';
+export { default as generateAttestModuleTests } from './attest-module';
+export { default as generateBlobModuleTests } from './blob-module';
+export { default as generateDIDModuleTests } from './did-module';
+export { default as generateOffchainSignatureModuleTests } from './offchain-signatures-module';
+export { default as generateStatusListCredentialModuleTests } from './status-list-credential-module';