+++ packages/credential-sdk/src/types/did/document/index.js
@@ -7,41 +7,41 @@ import {
   TypedUUID,
   option,
   withFrom,
-} from "../../generic";
+} from '../../generic';
 import {
   NamespaceDid,
   CheqdNamespaceDid,
   CheqdTestnetDid,
   CheqdMainnetDid,
-} from "../onchain/typed-did";
-import { DidKey, DidKeys } from "../onchain/did-key";
-import { VerificationRelationship } from "../onchain/verification-relationship";
+} from '../onchain/typed-did';
+import { DidKey, DidKeys } from '../onchain/did-key';
+import { VerificationRelationship } from '../onchain/verification-relationship';
 import {
   Service,
   CheqdService,
   ServiceEndpointId,
   CheqdTestnetService,
   CheqdMainnetService,
-} from "./service-endpoint";
+} from './service-endpoint';
 import {
   VerificationMethod,
   CheqdVerificationMethod,
   CheqdTestnetVerificationMethod,
   CheqdMainnetVerificationMethod,
-} from "./verification-method";
+} from './verification-method';
 import {
   CheqdVerificationMethodRefOrCheqdVerificationMethod,
   CheqdVerificationMethodRefOrCheqdMainnetVerificationMethod,
   CheqdVerificationMethodRefOrCheqdTestnetVerificationMethod,
-} from "./verification-method-ref-or-cheqd-verification-method";
+} from './verification-method-ref-or-cheqd-verification-method';
 import {
   VerificationMethodRef,
   CheqdVerificationMethodRef,
   CheqdTestnetVerificationMethodRef,
   CheqdMainnetVerificationMethodRef,
-} from "./verification-method-ref";
-import { ATTESTS_IRI, CONTEXT_URI } from "./const";
-import { ensureEqualToOrPrototypeOf } from "../../../utils";
+} from './verification-method-ref';
+import { ATTESTS_IRI, CONTEXT_URI } from './const';
+import { ensureEqualToOrPrototypeOf } from '../../../utils';
 
 class Context extends TypedArray {
   static Class = TypedString;
@@ -139,11 +139,10 @@ class CheqdMainnetAssertionMethod extends TypedArray {
 
 export class DIDDocument extends withFrom(
   TypedStruct,
-  (value, from) =>
-    from(value instanceof CheqdDIDDocument ? value.toDIDDocument() : value) // eslint-disable-line no-use-before-define
+  (value, from) => from(value instanceof CheqdDIDDocument ? value.toDIDDocument() : value), // eslint-disable-line no-use-before-define
 ) {
   static Classes = {
-    "@context": Context,
+    '@context': Context,
     id: ID,
     alsoKnownAs: option(AlsoKnownAs),
     controller: Controllers,
@@ -167,7 +166,7 @@ export class DIDDocument extends withFrom(
       alsoKnownAs = [],
       capabilityDelegation = [],
       attests = null,
-    } = {}
+    } = {},
   ) {
     const doc = new this(
       context,
@@ -181,7 +180,7 @@ export class DIDDocument extends withFrom(
       [],
       [],
       capabilityDelegation,
-      attests
+      attests,
     );
 
     let keyIdx = 0;
@@ -195,7 +194,7 @@ export class DIDDocument extends withFrom(
         } catch {
           return [ServiceEndpointId.from([did, ref]), serviceEndpoint];
         }
-      }
+      },
     );
 
     for (const [keyRef, didKey] of keysWithRefs) {
@@ -258,9 +257,7 @@ export class DIDDocument extends withFrom(
 
   removeKey(keyRef) {
     const ref = VerificationMethodRef.from(keyRef);
-    const keyIndex = this.verificationMethod.findIndex((method) =>
-      method.id.eq(ref)
-    );
+    const keyIndex = this.verificationMethod.findIndex((method) => method.id.eq(ref));
 
     // eslint-disable-next-line no-bitwise
     if (~keyIndex) {
@@ -306,8 +303,8 @@ export class DIDDocument extends withFrom(
     return TypedNumber.from(
       [...this.verificationMethod].reduce(
         (max, { id: { index } }) => Math.max(max, index ?? 0),
-        0
-      )
+        0,
+      ),
     ).inc();
   }
 
@@ -372,7 +369,7 @@ export class DIDDocument extends withFrom(
   // eslint-disable-next-line no-use-before-define
   toCheqd(Class = CheqdDIDDocument, versionId = TypedUUID.random()) {
     const {
-      "@context": context,
+      '@context': context,
       id,
       alsoKnownAs,
       controller,
@@ -398,7 +395,7 @@ export class DIDDocument extends withFrom(
       keyAgreement,
       capabilityInvocation,
       capabilityDelegation,
-      versionId
+      versionId,
     );
   }
 }
@@ -425,17 +422,14 @@ export class CheqdDIDDocument extends TypedStruct {
     const { verificationMethod } = this;
 
     this.verificationMethod = verificationMethod.filter(
-      (verMethod) =>
-        !verMethod.isOffchain() &&
-        !verMethod.verificationMaterial.bytes.every((item) => !item)
+      (verMethod) => !verMethod.isOffchain()
+        && !verMethod.verificationMaterial.bytes.every((item) => !item),
     );
-    const offchainVerMethod = verificationMethod.filter((verMethod) =>
-      verMethod.isOffchain()
-    );
+    const offchainVerMethod = verificationMethod.filter((verMethod) => verMethod.isOffchain());
 
     this.assertionMethod = [
       ...this.assertionMethod.filter(
-        (ref) => !offchainVerMethod.some((verMethod) => verMethod.id.eq(ref))
+        (ref) => !offchainVerMethod.some((verMethod) => verMethod.id.eq(ref)),
       ),
       ...offchainVerMethod,
     ];
@@ -457,14 +451,14 @@ export class CheqdDIDDocument extends TypedStruct {
     } = this;
 
     const assertionMethodOffchainKeys = [...assertionMethod].filter(
-      (keyRefOrKey) => keyRefOrKey.id
+      (keyRefOrKey) => keyRefOrKey.id,
     );
     const verificationMethodWithOffchainKeys = [
       ...verificationMethod,
       ...assertionMethodOffchainKeys,
     ];
     const assertionMethodOnlyRefs = [...assertionMethod].map(
-      (keyRefOrKey) => keyRefOrKey.id ?? keyRefOrKey
+      (keyRefOrKey) => keyRefOrKey.id ?? keyRefOrKey,
     );
 
     return new DIDDocument(
@@ -479,7 +473,7 @@ export class CheqdDIDDocument extends TypedStruct {
       keyAgreement,
       capabilityInvocation,
       capabilityDelegation,
-      null
+      null,
     );
   }
 }
@@ -518,9 +512,9 @@ export class CheqdMainnetDIDDocument extends CheqdDIDDocument {
   };
 }
 
-export * from "./const";
-export * from "./service-endpoint";
-export * from "./verification-method";
-export * from "./verification-method-type";
-export { default as IdentRef } from "./ident-ref";
+export * from './const';
+export * from './service-endpoint';
+export * from './verification-method';
+export * from './verification-method-type';
+export { default as IdentRef } from './ident-ref';
 export { VerificationMethodRef };