+++ packages/credential-sdk/src/types/offchain-signatures/public-keys/index.js
@@ -1,24 +1,26 @@
-import { TypedEnum, option, withProp } from "../../generic";
+import { TypedEnum, option, withProp } from '../../generic';
 import {
   Bls12381BBS23DockVerKeyName,
   Bls12381BBSDockVerKeyName,
   Bls12381PSDockVerKeyName,
-} from "../../../vc/crypto";
+} from '../../../vc/crypto';
 import {
   BBSPublicKeyValue,
   BBSPlusPublicKeyValue,
   PSPublicKeyValue,
   BBDT16PublicKeyValue,
-} from "./value";
+} from './value';
 import {
   CheqdMainnetOffchainSignatureParamsRef,
   CheqdOffchainSignatureParamsRef,
   CheqdTestnetOffchainSignatureParamsRef,
   DockOffchainSignatureParamsRef,
-} from "./ref";
-import { Bls12381BBDT16DockVerKeyName } from "../../../vc/crypto/constants";
-import { BBDT16Params, BBSParams, BBSPlusParams, PSParams } from "../params";
-import { maybeToJSONString } from "../../../utils";
+} from './ref';
+import { Bls12381BBDT16DockVerKeyName } from '../../../vc/crypto/constants';
+import {
+  BBDT16Params, BBSParams, BBSPlusParams, PSParams,
+} from '../params';
+import { maybeToJSONString } from '../../../utils';
 
 export class OffchainSignaturePublicKey extends TypedEnum {
   get bytes() {
@@ -44,8 +46,8 @@ export class OffchainSignaturePublicKey extends TypedEnum {
   setParams(params) {
     const WithParams = withProp(
       this.constructor,
-      "params",
-      option(this.constructor.Params)
+      'params',
+      option(this.constructor.Params),
     );
 
     const withParams = WithParams.from(this);
@@ -62,8 +64,8 @@ export class OffchainSignaturePublicKey extends TypedEnum {
       if (params == null) {
         throw new Error(
           `Parameters with reference (${maybeToJSONString(
-            this.paramsRef
-          )}) not found on chain`
+            this.paramsRef,
+          )}) not found on chain`,
         );
       }
     }
@@ -76,7 +78,7 @@ export class BBSPublicKey extends OffchainSignaturePublicKey {
 
   static Params = BBSParams;
 
-  static Type = "bbs";
+  static Type = 'bbs';
 
   static VerKeyType = Bls12381BBS23DockVerKeyName;
 }
@@ -85,7 +87,7 @@ export class BBSPlusPublicKey extends OffchainSignaturePublicKey {
 
   static Params = BBSPlusParams;
 
-  static Type = "bbsPlus";
+  static Type = 'bbsPlus';
 
   static VerKeyType = Bls12381BBSDockVerKeyName;
 }
@@ -94,7 +96,7 @@ export class PSPublicKey extends OffchainSignaturePublicKey {
 
   static Params = PSParams;
 
-  static Type = "ps";
+  static Type = 'ps';
 
   static VerKeyType = Bls12381PSDockVerKeyName;
 }
@@ -103,7 +105,7 @@ export class BBDT16PublicKey extends OffchainSignaturePublicKey {
 
   static Params = BBDT16Params;
 
-  static Type = "bbdt16";
+  static Type = 'bbdt16';
 
   static VerKeyType = Bls12381BBDT16DockVerKeyName;
 }
@@ -112,32 +114,32 @@ OffchainSignaturePublicKey.bindVariants(
   BBSPublicKey,
   BBSPlusPublicKey,
   PSPublicKey,
-  BBDT16PublicKey
+  BBDT16PublicKey,
 );
 
 export class DockOffchainSignaturePublicKey extends withProp(
   OffchainSignaturePublicKey,
-  "paramsRef",
-  option(DockOffchainSignatureParamsRef)
+  'paramsRef',
+  option(DockOffchainSignatureParamsRef),
 ) {}
 
 export class CheqdOffchainSignaturePublicKey extends withProp(
   OffchainSignaturePublicKey,
-  "paramsRef",
-  option(CheqdOffchainSignatureParamsRef)
+  'paramsRef',
+  option(CheqdOffchainSignatureParamsRef),
 ) {}
 
 export class CheqdTestnetOffchainSignaturePublicKey extends withProp(
   OffchainSignaturePublicKey,
-  "paramsRef",
-  option(CheqdTestnetOffchainSignatureParamsRef)
+  'paramsRef',
+  option(CheqdTestnetOffchainSignatureParamsRef),
 ) {}
 export class CheqdMainnetOffchainSignaturePublicKey extends withProp(
   OffchainSignaturePublicKey,
-  "paramsRef",
-  option(CheqdMainnetOffchainSignatureParamsRef)
+  'paramsRef',
+  option(CheqdMainnetOffchainSignatureParamsRef),
 ) {}
 
-export * from "./value";
-export * from "./ref";
-export * from "./id";
+export * from './value';
+export * from './ref';
+export * from './id';