+++ packages/dock-blockchain-modules/src/did/internal.js
@@ -9,19 +9,19 @@ import {
   DidKeys,
   Controllers,
   StoredDidDetails,
-} from "@docknetwork/credential-sdk/types";
+} from '@docknetwork/credential-sdk/types';
 import {
   option,
   TypedNumber,
   TypedTuple,
-} from "@docknetwork/credential-sdk/types/generic";
+} from '@docknetwork/credential-sdk/types/generic';
 import {
   NoDIDError,
   NoOffchainDIDError,
   NoOnchainDIDError,
-} from "@docknetwork/credential-sdk/modules/abstract/did";
-import { maybeToHuman } from "@docknetwork/credential-sdk/utils";
-import { DidMethodKeyDetails } from "@docknetwork/credential-sdk/types/did/onchain";
+} from '@docknetwork/credential-sdk/modules/abstract/did';
+import { maybeToHuman } from '@docknetwork/credential-sdk/utils';
+import { DidMethodKeyDetails } from '@docknetwork/credential-sdk/types/did/onchain';
 import {
   AddServiceEndpoint,
   AddKeys,
@@ -30,8 +30,8 @@ import {
   RemoveControllers,
   RemoveServiceEndpoint,
   RemoveOnchainDid,
-} from "./actions";
-import createInternalDockModule from "../common/create-internal-dock-module";
+} from './actions';
+import createInternalDockModule from '../common/create-internal-dock-module';
 
 const didMethods = {
   addKeys: (keys, targetDid, _, nonce) => {
@@ -46,13 +46,12 @@ const didMethods = {
     return new AddControllers(did, controllers, nonce);
   },
 
-  addServiceEndpoint: (id, types, origins, targetDid, _, nonce) =>
-    new AddServiceEndpoint(
-      DockDid.from(targetDid).asDid,
-      id,
-      new ServiceEndpoint(types, origins),
-      nonce
-    ),
+  addServiceEndpoint: (id, types, origins, targetDid, _, nonce) => new AddServiceEndpoint(
+    DockDid.from(targetDid).asDid,
+    id,
+    new ServiceEndpoint(types, origins),
+    nonce,
+  ),
 
   removeKeys: (keys, targetDid, _, nonce) => {
     const did = DockDid.from(targetDid).asDid;
@@ -118,10 +117,10 @@ export default class DockDIDModuleInternal extends createInternalDockModule({
   didMethods,
   accountMethods,
 }) {
-  static Prop = "didModule";
+  static Prop = 'didModule';
 
   static MethodNameOverrides = {
-    removeOnchainDid: "DidRemoval",
+    removeOnchainDid: 'DidRemoval',
   };
 
   /**
@@ -198,7 +197,7 @@ export default class DockDIDModuleInternal extends createInternalDockModule({
       controllers,
       targetDid,
       didKeypair,
-      params
+      params,
     );
   }
 
@@ -218,7 +217,7 @@ export default class DockDIDModuleInternal extends createInternalDockModule({
     origins,
     targetDid,
     didKeypair,
-    params
+    params,
   ) {
     return await this.send.addServiceEndpoint(
       endpointId,
@@ -226,7 +225,7 @@ export default class DockDIDModuleInternal extends createInternalDockModule({
       origins,
       targetDid,
       didKeypair,
-      params
+      params,
     );
   }
 
@@ -255,7 +254,7 @@ export default class DockDIDModuleInternal extends createInternalDockModule({
       controllers,
       targetDid,
       didKeypair,
-      params
+      params,
     );
   }
 
@@ -271,7 +270,7 @@ export default class DockDIDModuleInternal extends createInternalDockModule({
     return await this.send.removeServiceEndpoint(
       endpointId,
       didKeypair,
-      params
+      params,
     );
   }
 
@@ -298,7 +297,7 @@ export default class DockDIDModuleInternal extends createInternalDockModule({
 
     return (
       option(TypedNumber).from(
-        await this.query.didControllers(controlledDid, controllerDid)
+        await this.query.didControllers(controlledDid, controllerDid),
       ) != null
     );
   }
@@ -313,12 +312,12 @@ export default class DockDIDModuleInternal extends createInternalDockModule({
     const [owner, id] = ServiceEndpointId.from(endpointId);
 
     const endpoint = option(ServiceEndpoint).from(
-      await this.query.didServiceEndpoints(owner.asDock, id)
+      await this.query.didServiceEndpoints(owner.asDock, id),
     );
 
     if (endpoint == null) {
       throw new Error(
-        `No service endpoint found for did ${owner} and with id ${endpointId}`
+        `No service endpoint found for did ${owner} and with id ${endpointId}`,
       );
     }
 
@@ -352,7 +351,7 @@ export default class DockDIDModuleInternal extends createInternalDockModule({
   async getOnchainDidDetail(rawDid) {
     const did = DockDid.from(rawDid);
     const resp = option(StoredDidDetails).from(
-      await this.query.dids(did.asDid)
+      await this.query.dids(did.asDid),
     );
     if (resp == null) {
       throw new NoDIDError(did);
@@ -367,7 +366,7 @@ export default class DockDIDModuleInternal extends createInternalDockModule({
   async getDidMethodKeyDetail(did) {
     const didKey = DidMethodKey.from(did);
     const resp = option(DidMethodKeyDetails).from(
-      await this.query.didMethodKeys(didKey.asDidMethodKey)
+      await this.query.didMethodKeys(didKey.asDidMethodKey),
     );
     if (resp == null) {
       throw new NoDIDError(didKey);
@@ -384,7 +383,7 @@ export default class DockDIDModuleInternal extends createInternalDockModule({
   async getOffchainDidDetail(didIdentifier) {
     const did = DockDid.from(didIdentifier);
     const resp = option(StoredDidDetails).from(
-      await this.query.dids(did.asDid)
+      await this.query.dids(did.asDid),
     );
     if (resp == null) {
       throw new NoDIDError(did);
@@ -425,11 +424,11 @@ export default class DockDIDModuleInternal extends createInternalDockModule({
 
   async controllers(did) {
     const controllers = await this.query.didControllers.entries(
-      DockDid.from(did).asDid
+      DockDid.from(did).asDid,
     );
 
     return Controllers.from(
-      controllers.map(([entry]) => maybeToHuman(entry)[1])
+      controllers.map(([entry]) => maybeToHuman(entry)[1]),
     );
   }
 
@@ -441,13 +440,11 @@ export default class DockDIDModuleInternal extends createInternalDockModule({
 
     return new ServiceEndpoints(
       (await this.query.didServiceEndpoints.entries(typedDid.asDid))
-        .map(([key, value]) =>
-          ServiceEndpointEntry.from([
-            [typedDid, maybeToHuman(key)[1]],
-            maybeToHuman(value),
-          ])
-        )
-        .filter(([_, sp]) => sp)
+        .map(([key, value]) => ServiceEndpointEntry.from([
+          [typedDid, maybeToHuman(key)[1]],
+          maybeToHuman(value),
+        ]))
+        .filter(([_, sp]) => sp),
     );
   }
 
@@ -458,11 +455,9 @@ export default class DockDIDModuleInternal extends createInternalDockModule({
 
     return new DidKeys(
       (await this.query.didKeys.entries(DockDid.from(did).asDid))
-        .map(([key, value]) =>
-          DidKeyWithId.from([maybeToHuman(key)[1], maybeToHuman(value)])
-        )
+        .map(([key, value]) => DidKeyWithId.from([maybeToHuman(key)[1], maybeToHuman(value)]))
         .map(([id, key]) => [[did, id], key])
-        .filter(([_, pk]) => pk)
+        .filter(([_, pk]) => pk),
     );
   }
 }