+++ packages/credential-sdk/CHANGELOG.md
@@ -4,6 +4,12 @@
 
 ### Minor Changes
 
+- Add tests for all types of identifiers that can be received by the modules, use replacement resolver for DID documents
+
+## 0.39.0
+
+### Minor Changes
+
 - Use parsed accumulator identifier in `createAccumulator`/`updateAccumulator` calls
 
 ## 0.38.0